+++ yo/app/scripts/services/tc-user.service.js
@@ -233,6 +233,47 @@
                     return this.setDownloadStatus("" + id, status, {});
                 }
             });
+            
+            this.getDownloadTypeStatus = helpers.overload({
+            	/**
+            	 * @method
+            	 * @name User#getDownloadTypeStatus
+            	 * @param {string} downloadType download type as defined in topcat.json (e.g. 'globus')
+            	 * @param {object} options {@link https://docs.angularjs.org/api/ng/service/$http#usage|as specified in the Angular documentation}
+            	 * @return {Promise<object>} a deferred object containing disabled flag and string message
+            	 */
+            	'string,object' : function(downloadType,options){
+            		return this.get('downloadType/' + downloadType + '/status', {
+                        facilityName: facility.config().name,
+            			sessionId: facility.icat().session().sessionId
+            		}, options).then(function(status){
+            			return status;
+            		}, function(response){
+            			// error case - log but treat as not disabled
+            			console.log('user.getDownloadTypeStatus failed - treat as not disabled');
+            			return {disabled:false,message:""};
+            		})
+            	},
+            	/**
+            	 * @method
+            	 * @name User#getDownloadTypeStatus
+            	 * @param {string} downloadType download type as defined in topcat.json (e.g. 'globus')
+                 * @param  {Promise} timeout if resolved will cancel the request
+            	 * @return {Promise<object>} a deferred object containing disabled flag and string message
+            	 */
+            	'string,promise' : function(downloadType,timeout){
+            		return this.getDownloadTypeStatus(downloadType,{timeout: timeout});
+            	},
+            	/**
+            	 * @method
+            	 * @name User#getDownloadTypeStatus
+            	 * @param {string} downloadType download type as defined in topcat.json (e.g. 'globus')
+            	 * @return {Promise<object>} a deferred object containing disabled flag and string message
+            	 */
+            	'string': function(downloadType){
+            		return this.getDownloadTypeStatus(downloadType,{});
+            	}
+            });
 
             var cartCache;