+++ src/site/markdown/release-notes.md
@@ -8,6 +8,7 @@
   * Remove DOI and some login redirections from browser history (issue #424)
   * Allow ids.timeout to be in seconds or minutes as well as milliseconds (issue #349)
   * Add optional displayName and description to downloadTransportTypes (issue #427)
+  * Add API to disable/enable download transport types (issue #434)
   
 ## 2.4.3 (25th Mar 2019)